Use add_llvm_executable instead of add_llvm_tool.
add_llvm_tool has some extra configurability that can exclude it from the ALL
target.  We don't really want to include all of the llvm configuration
machinery in llvm/CMakeLists.txt, but withouht this machinery, the default
is to exclude from the ALL target.  This change results in standalone-opt
getting built even if the tests are not run.